Cartpole parameters handled by separate class
